requirements.txt.txt
python-telegram-bot==13.15
schedule==1.2.0
